Refactor ArtworkListContainer to use custom hook for filtered artworks
